Solution to Calculate Average Marks of Students

+++ Calculate Average Marks of Students@@67877d61dc107486ad5c2743/code.c
@@ -19,6 +19,6 @@ int main(){
         
         sum+=std[i].marks;
     }
-    printf("Average Marks: %.2f",sum/(float)n);
+    printf("Average Marks: %.2f",sum/n);
     return 0;
 }
